backned/requirements.txt
flask
pymongo
bson
python-dotenv
more-itertools
requests
nltk
